fix: change tiles count from int to long (#38)
* fix: change tiles count from int to long

* fix: fix cr comments

* fix: fix cr comments

* fix: add cr review test

+++ MergerCli/Process.cs
@@ -20,8 +20,8 @@ namespace MergerCli
         {
             batchStatusManager.InitilaizeLayer(newData.Path);
             List<Tile> tiles = new List<Tile>(batchSize);
-            int totalTileCount = newData.TileCount();
-            int tileProgressCount = 0;
+            long totalTileCount = newData.TileCount();
+            long tileProgressCount = 0;
 
             string? resumeBatchIdentifier = batchStatusManager.GetLayerBatchIdentifier(newData.Path);
             if (resumeBatchIdentifier != null)
@@ -78,8 +78,8 @@ namespace MergerCli
             List<Tile> newTiles;
             bool hasSameTiles = true;
 
-            int totalTileCount = newData.TileCount();
-            int tilesChecked = 0;
+            long totalTileCount = newData.TileCount();
+            long tilesChecked = 0;
             this._logger.LogInformation($"Base tile Count: {baseData.TileCount()}, New tile count: {totalTileCount}");
 
             do
@@ -88,7 +88,6 @@ namespace MergerCli
 
                 int baseMatchCount = 0;
                 int newTileCount = 0;
-
                 for (int i = 0; i < newTiles.Count; i++)
                 {
                     Tile newTile = newTiles[i];
