+++ MergerLogic/Utils/IGpkgUtils.cs
@@ -6,10 +6,10 @@ namespace MergerLogic.Utils
     public interface IGpkgUtils : IDataUtils
     {
         void CreateTileIndex();
-        List<Tile> GetBatch(int batchSize, int offset);
+        List<Tile> GetBatch(int batchSize, long offset);
         Extent GetExtent();
         Tile GetLastTile(int[] coords, Coord baseCoords);
-        int GetTileCount();
+        long GetTileCount();
         void InsertTiles(IEnumerable<Tile> tiles);
         void UpdateExtent(Extent extent);
         void Vacuum();
